@@ -6,17 +6,18 @@
   "main": "server.js",
   "scripts": {
     "postinstall": "cd client && npm install",
-    "build": "cd client && npm run build && npm run build-storybook",
+    "build:server": "tsc",
+    "build": "npm run build:server && cd client && npm run build && npm run build-storybook",
     "test:watch": "cd client && npm run test:watch",
     "test": "npm run lint && npm run prettier && cd client && npm test",
-    "lint": "eslint . --ext .js",
+    "lint": "eslint . --ext .{js,jsx,ts,tsx}",
     "prettify": "prettier --write .",
     "dev": "concurrently \"npm run server\" \"npm run client\"",
     "client": "cd client && npm start",
     "prettier": "prettier --check \"**/*.{js,jsx,ts,tsx,md,mdx,html,css,json}\"",
-    "server": "nodemon server.js",
+    "server": "ts-node-dev server.ts",
     "storybook": "cd client && npm run storybook",
-    "start": "node server.js"
+    "start": "node dist/server.js"
   },
   "repository": {
     "type": "git",
@@ -32,14 +33,20 @@
   "dependencies": {
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "mongodb": "^3.6.3"
+    "mongodb": "^3.6.3",
+    "ts-node-dev": "^1.0.0",
+    "typescript": "^4.1.2"
   },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.8.1",
-    "@typescript-eslint/parser": "^4.8.1",
+    "@types/express": "^4.17.9",
+    "@types/jest": "^26.0.15",
+    "@types/mongodb": "^3.5.34",
+    "@types/node": "^14.14.9",
+    "@typescript-eslint/eslint-plugin": "^4.8.2",
+    "@typescript-eslint/parser": "^4.8.2",
     "babel-eslint": "^10.1.0",
     "concurrently": "^5.3.0",
-    "eslint": "^7.13.0",
+    "eslint": "^7.14.0",
     "eslint-config-prettier": "^6.15.0",
     "eslint-config-react-app": "^6.0.0",
     "eslint-plugin-flowtype": "^5.2.0",
@@ -49,9 +56,7 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "husky": "^4.3.0",
     "lint-staged": "^10.5.1",
-    "nodemon": "^2.0.6",
-    "prettier": "2.1.2",
-    "typescript": "^4.0.5"
+    "prettier": "2.1.2"
   },
   "husky": {
     "hooks": {
@@ -60,7 +65,7 @@
     }
   },
   "lint-staged": {
-    "*.js": "eslint --cache --fix",
-    "*.{js,css,md}": "prettier --write"
+    "*.{js,jsx,ts,tsx}": "eslint --cache --fix",
+    "*.{js,jsx,ts,tsx,css,md}": "prettier --write"
   }
 }
